+++ CHANGELOG.md
@@ -17,6 +17,35 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Chore 👨‍💻 👩‍💻
 
+## [1.58.1] - 2020-10-02
+
+### Added 🚀
+
+### Changed
+
+### Removed 🗑
+
+### Fixed 🐞
+
+-   Showing wrong edges when hovering a building after selecting one #1137
+
+### Chore 👨‍💻 👩‍💻
+
+## [1.58.0] - 2020-10-02
+
+### Added 🚀
+
+-   Add active color metric to the top of the legend panel #1278
+-   SourceCodeParser: Java 14 Support #1277
+
+### Changed
+
+### Removed 🗑
+
+### Fixed 🐞
+
+### Chore 👨‍💻 👩‍💻
+
 ## [1.57.4] - 2020-09-25
 
 ### Added 🚀